Make API more reusable
Since both the ground() and get_models() function rely on a static instance of the GrounderInstance, I moved the implementations of these functions into the class itself. Later, this made it possible to just call these methods on the static instance of the GrounderInstance from the api-level functions.

This will enable a user to make their own GrounderInstance if they would like