Include package data
right now, if you do `from gilda.resources import MESH_MAPPINGS_PATH`, there isn't actually a file there. If you set `include_package_data=True` in the setup configuration, this file should get propagated through to the package. I'm pretty sure this is the case, and if not, you just need to add something to the MANIFEST.in file.