Make the web dependencies optional
There are several situations in which installing the web dependencies for Gilda by default is inconvenient. The largest are due to conflicts in versions of werkzeug (which is pinned in gilda) and conflicts between flask-bootstrap (the one gilda uses, but it's abandoned) and bootstrap-flask (fork other former and is currently maintained)

As it is, it's not possible to easily have gilda simultaneously installed with biomappings, and this is a concrete issue for Jeremy T who wants to contribute some curations (without him doing some more manual dependency management)

This PR simply makes the web dependencies optional. It updates the dockerfile and tests to make sure these are still installed where necessary, and updates relevant documentation accordingly